requirements.txt
appdirs==1.4.4
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
black==19.10b0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.7
cleo==0.8.1
click==7.1.2
clikit==0.6.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==3.3.2
cssselect==1.1.0
decorator==4.4.1
demjson==2.2.4
distlib==0.3.3
Django==2.2.13
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-guardian==2.0.0
django-model-utils==4.2.0
djangorestframework==3.9.4
dodgy==0.2.1
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-PyMongo==2.3.0
html5lib==1.1
hyperlink==18.0.0
idna==2.9
incremental==17.5.0
iniconfig==1.1.1
ipdb==0.11
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.16.0
jeepney==0.7.1
Jinja2==2.11.2
keyring==21.8.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
Markdown==3.1.1
markdown2==2.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==2.0.0
msgpack==1.0.2
netaddr==0.7.20
packaging==20.9
paramiko==2.7.1
parsel==1.5.1
parso==0.7.0
pastel==0.2.1
pathspec==0.8.0
pep8-naming==0.4.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.6.2
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
poetry==1.1.11
poetry-core==1.0.7
prompt-toolkit==3.0.4
prospector==1.2.0
Protego==0.1.16
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pydocstyle==5.0.2
pyflakes==2.1.0
Pygments==2.7.4
PyHamcrest==2.0.2
pylev==1.4.0
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pymongo==3.10.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyquery==1.4.3
pytest==6.2.5
python-decouple==3.5
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
readtime==1.1.1
regex==2020.5.14
requests==2.23.0
requests-toolbelt==0.9.1
requirements-detector==0.7
scapy==2.4.3
Scrapy==1.8.1
scrapyrt==0.11.0
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.4.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.2.1
sqlparse==0.3.1
toml==0.10.1
tomlkit==0.7.2
traitlets==4.3.2
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.8.1
virtualenv-clone==0.5.4
w3lib==1.19.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
zope.interface==4.7.1
